requirements.txt
fastapi==0.115.0
Jinja2==3.1.4
mss==9.0.1
Pillow==10.3.0
pynput==1.7.6
pyautogui==0.9.54
pygetwindow==0.0.9
pystray==0.19.5
requests==2.32.3
uvicorn==0.30.0
